+++ src/Twilio/Rest/Conversations/V1/ConversationOptions.php
@@ -18,6 +18,8 @@ use Twilio\Values;
 abstract class ConversationOptions {
     /**
      * @param string $friendlyName The human-readable name of this conversation.
+     * @param string $uniqueName An application-defined string that uniquely
+     *                           identifies the resource
      * @param \DateTime $dateCreated The date that this resource was created.
      * @param \DateTime $dateUpdated The date that this resource was last updated.
      * @param string $messagingServiceSid The unique id of the SMS Service this
@@ -33,8 +35,8 @@ abstract class ConversationOptions {
      *                                      request header
      * @return CreateConversationOptions Options builder
      */
-    public static function create(string $friendlyName = Values::NONE, \DateTime $dateCreated = Values::NONE, \DateTime $dateUpdated = Values::NONE, string $messagingServiceSid = Values::NONE, string $attributes = Values::NONE, string $state = Values::NONE, string $timersInactive = Values::NONE, string $timersClosed = Values::NONE, string $xTwilioWebhookEnabled = Values::NONE): CreateConversationOptions {
-        return new CreateConversationOptions($friendlyName, $dateCreated, $dateUpdated, $messagingServiceSid, $attributes, $state, $timersInactive, $timersClosed, $xTwilioWebhookEnabled);
+    public static function create(string $friendlyName = Values::NONE, string $uniqueName = Values::NONE, \DateTime $dateCreated = Values::NONE, \DateTime $dateUpdated = Values::NONE, string $messagingServiceSid = Values::NONE, string $attributes = Values::NONE, string $state = Values::NONE, string $timersInactive = Values::NONE, string $timersClosed = Values::NONE, string $xTwilioWebhookEnabled = Values::NONE): CreateConversationOptions {
+        return new CreateConversationOptions($friendlyName, $uniqueName, $dateCreated, $dateUpdated, $messagingServiceSid, $attributes, $state, $timersInactive, $timersClosed, $xTwilioWebhookEnabled);
     }
 
     /**
@@ -71,6 +73,8 @@ abstract class ConversationOptions {
 class CreateConversationOptions extends Options {
     /**
      * @param string $friendlyName The human-readable name of this conversation.
+     * @param string $uniqueName An application-defined string that uniquely
+     *                           identifies the resource
      * @param \DateTime $dateCreated The date that this resource was created.
      * @param \DateTime $dateUpdated The date that this resource was last updated.
      * @param string $messagingServiceSid The unique id of the SMS Service this
@@ -85,8 +89,9 @@ class CreateConversationOptions extends Options {
      * @param string $xTwilioWebhookEnabled The X-Twilio-Webhook-Enabled HTTP
      *                                      request header
      */
-    public function __construct(string $friendlyName = Values::NONE, \DateTime $dateCreated = Values::NONE, \DateTime $dateUpdated = Values::NONE, string $messagingServiceSid = Values::NONE, string $attributes = Values::NONE, string $state = Values::NONE, string $timersInactive = Values::NONE, string $timersClosed = Values::NONE, string $xTwilioWebhookEnabled = Values::NONE) {
+    public function __construct(string $friendlyName = Values::NONE, string $uniqueName = Values::NONE, \DateTime $dateCreated = Values::NONE, \DateTime $dateUpdated = Values::NONE, string $messagingServiceSid = Values::NONE, string $attributes = Values::NONE, string $state = Values::NONE, string $timersInactive = Values::NONE, string $timersClosed = Values::NONE, string $xTwilioWebhookEnabled = Values::NONE) {
         $this->options['friendlyName'] = $friendlyName;
+        $this->options['uniqueName'] = $uniqueName;
         $this->options['dateCreated'] = $dateCreated;
         $this->options['dateUpdated'] = $dateUpdated;
         $this->options['messagingServiceSid'] = $messagingServiceSid;
@@ -105,6 +110,18 @@ class CreateConversationOptions extends Options {
      */
     public function setFriendlyName(string $friendlyName): self {
         $this->options['friendlyName'] = $friendlyName;
+        return $this;
+    }
+
+    /**
+     * An application-defined string that uniquely identifies the resource. It can be used to address the resource in place of the resource's `sid` in the URL.
+     *
+     * @param string $uniqueName An application-defined string that uniquely
+     *                           identifies the resource
+     * @return $this Fluent Builder
+     */
+    public function setUniqueName(string $uniqueName): self {
+        $this->options['uniqueName'] = $uniqueName;
         return $this;
     }