+++ src/Twilio/Rest/Serverless/V1/Service/Environment/DeploymentOptions.php
@@ -17,7 +17,7 @@ use Twilio\Values;
  */
 abstract class DeploymentOptions {
     /**
-     * @param string $buildSid The SID of the build for the deployment
+     * @param string $buildSid The SID of the Build for the Deployment
      * @return CreateDeploymentOptions Options builder
      */
     public static function create(string $buildSid = Values::NONE): CreateDeploymentOptions {
@@ -27,16 +27,16 @@ abstract class DeploymentOptions {
 
 class CreateDeploymentOptions extends Options {
     /**
-     * @param string $buildSid The SID of the build for the deployment
+     * @param string $buildSid The SID of the Build for the Deployment
      */
     public function __construct(string $buildSid = Values::NONE) {
         $this->options['buildSid'] = $buildSid;
     }
 
     /**
-     * The SID of the build for the deployment.
+     * The SID of the Build for the Deployment.
      *
-     * @param string $buildSid The SID of the build for the deployment
+     * @param string $buildSid The SID of the Build for the Deployment
      * @return $this Fluent Builder
      */
     public function setBuildSid(string $buildSid): self {