+++ src/Twilio/Rest/Taskrouter/V1/Workspace/ActivityOptions.php
@@ -129,7 +129,7 @@ class CreateActivityOptions extends Options {
     }
 
     /**
-     * Whether the Worker should be eligible to receive a Task when it occupies the Activity. A value of `true`, `1`, or `yes` specifies the Activity is available. All other values specify that it is not.
+     * Whether the Worker should be eligible to receive a Task when it occupies the Activity. A value of `true`, `1`, or `yes` specifies the Activity is available. All other values specify that it is not. The value cannot be changed after the Activity is created.
      *
      * @param bool $available Whether the Worker should be eligible to receive a
      *                        Task when it occupies the Activity