+++ tests/Twilio/Integration/Conversations/V1/Conversation/ParticipantTest.php
@@ -46,7 +46,7 @@ class ParticipantTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -156,13 +156,13 @@ class ParticipantTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{}",
                 "messaging_binding": {
                     "type": "sms",
                     "address": "+15017122661"
                 },
-                "role_sid": "null",
+                "role_sid": null,
                 "date_created": "2020-07-01T22:18:37Z",
                 "date_updated": "2020-07-01T22:18:37Z",
                 "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -206,7 +206,7 @@ class ParticipantTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -312,7 +312,7 @@ class ParticipantTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -367,7 +367,7 @@ class ParticipantTest extends HolodeckTestCase {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "identity": "null",
+                        "identity": null,
                         "attributes": "{ \\"role\\": \\"driver\\" }",
                         "messaging_binding": {
                             "type": "sms",